requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# Security & Auth
cryptography==42.0.2
PyJWT==2.8.0
django-otp==1.3.0

# API & Utils
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.0
django-filter==23.5
django-storages==1.14.2

# Privacy & GDPR
django-gdpr-assist==1.4.0

# Environment
python-dotenv==1.0.0
django-environ==0.11.2
python-decouple==3.8

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development & Testing
pytest-django==4.7.0
factory-boy==3.3.0